+++ src/index.js
@@ -25,7 +25,7 @@ root.render(
       <Routes>
         <Route exact path='/' element={< App />}></Route>
         <Route exact path='/account' element={{/*link to account us*/}}></Route>
-        <Route exact path='/show/:showId' element={<ProgramPage/>}></Route>
+        <Route exact path='/show/:programId' element={<ProgramPage/>}></Route>
         <Route exact path='/programmering' element={<Programming />}></Route>
         <Route exact path='/ticket/:showId' element={<Ticket/>}></Route>
         <Route exact path='/begunstigers' element={{/*link to begunstigers us*/}}></Route>
